Added world meta data with gamestages, creation time, and last access time

+++ Assets/Scripts/Item/GameStage/TieredGameStage.cs
@@ -9,7 +9,7 @@ namespace Item.GameStage
         public Tier Tier;
         public override string GetGameStageId()
         {
-            return ((int)Tier).ToString();
+            return Tier.ToString();
         }
     }
 }
